# TC: Load existing node_modules from docker image