--- a/.teamcity/_self/projects/WebApp.kt
+++ b/.teamcity/_self/projects/WebApp.kt
@@ -44,7 +44,7 @@ object RunCalypsoE2eDesktopTests : BuildType({
 				export NODE_ENV="test"
 
 				# Install modules
-				yarn install
+				${_self.yarn_install_cmd}
 			"""
 			dockerImage = "%docker_image_e2e%"
 		}
@@ -210,7 +210,7 @@ object RunCalypsoE2eMobileTests : BuildType({
 				export NODE_ENV="test"
 
 				# Install modules
-				yarn install
+				${_self.yarn_install_cmd}
 			"""
 			dockerImage = "%docker_image_e2e%"
 		}
@@ -438,7 +438,7 @@ object RunAllUnitTests : BuildType({
 				export NODE_ENV="test"
 
 				# Install modules
-				yarn install
+				${_self.yarn_install_cmd}
 			"""
 		}
 		bashNodeScript {
@@ -651,7 +651,7 @@ object CheckCodeStyleBranch : BuildType({
 				export NODE_ENV="test"
 
 				# Install modules
-				yarn install
+				${_self.yarn_install_cmd}
 			"""
 		}
 		bashNodeScript {
@@ -744,7 +744,7 @@ object RunCalypsoPlaywrightE2eTests : BuildType({
 				export PLAYWRIGHT_BROWSERS_PATH=0
 
 				# Install modules
-				yarn install
+				${_self.yarn_install_cmd}
 
 				# Build package
 				yarn workspace @automattic/calypso-e2e build