--- a/.teamcity/_self/yarnInstall.kt
+++ b/.teamcity/_self/yarnInstall.kt
@@ -1,0 +1,17 @@
+package _self
+
+/**
+ * Use variable in scripts to take advantage of node_module caching.
+ */
+val yarn_install_cmd = """
+	# Load existing node_modules to reduce install time.
+	if [ -d /calypso/node_modules ] ; then
+		echo "Loading existing found node_modules..."
+		mv /calypso/node_modules ./node_modules
+	else
+		echo "No existing node_modules were found."
+	fi
+
+	# Install modules.
+	yarn install
+""".trimIndent()