#### Changes proposed in this Pull Request
Attempts loading the existing node_modules from the Docker image. (Added in #51421, [docker images rebuilt here](https://teamcity.a8c.com/buildConfiguration/calypso_BuildBaseImages/6038457?branch=%3Cdefault%3E&buildTypeTab=overview&mode=builds&buildTab=overview))

This should theoretically reduce `yarn install` time.

Initial results for ETK build:

| Action | Before | After |
| ------------- | ------------- | ------------- |
| Yarn install time: | 53s | 23.34s  |
| Prepare environment time: | 1:06 | 47s  |
| Total completion time: | 2:10 | 1:53 |

I believe this is reducing yarn install time by more than half, or about 30s. However, the typical `prepare environment` for all wpcom plugins is just above one minute. This only reduces it to <50s. In other words, I think the benefit could be bigger.

#### Testing instructions
1. ETK and other wpcom plugin builds should successfully complete
2. The prepare environment step should be faster than before. (Less than one minute, ideally)