@scinos 2021-03-26
What is the plan if, for whatever reason, a PR fails to build because a problem with the cached `node_mooules`?

We should have some kind of escape hatch to disable this feature in case something is wrong. Maybe a custom build parameter? I'm not sure.