@noahtallen 2021-03-26
I've added an escape hatch in 6992ef1. Basically, we should be able to set `use_cached_node_modules` to false as a param for any build, and it will not use the cache. 

Another option is to rebuild the base image without cached node modules